déplace les méthodes qui démarrent les containers pour les tests dans le package `test`
